requirements.txt
requests
pandas
python-dotenv
supabase
python-dateutil